+++ ChangeLog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Python class generation.
  - Perl class generation.
 
-## [0.3.0] - 2019-11-28
+## [0.3.0] - 2019-12-26
 ### Added
  - Examples of basic usage and features.
  - Option to give a class-level include list.
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    family of functions.
  - Classes can be explicitly specified as pointer or struct wrapping classes
    using the `type` property.
+ - Return values are not casted if the generated function and wrapped function
+   have the same return type.
 
 ### Fixed
  - `bin/wrapture` is now executable.